Split board modules from config
