+++ src/litexcnc/firmware/boards/etherbone/colorlight.py
@@ -1,17 +1,8 @@
-
-from typing import Literal
-
 # Imports for defining the board
 from liteeth.phy.ecp5rgmii import LiteEthPHYRGMII
 from litex.soc.integration.soc_core import *
 from litex_boards.targets.colorlight_5a_75x import _CRG
 from litex_boards.platforms import colorlight_5a_75b, colorlight_5a_75e
-
-# Imports from Pydantic to create the config
-from pydantic import Field
-
-# Local imports
-from . import Etherbone, EthPhy, EtherboneBoard
 
 
 class ColorLightBase(SoCMini):
@@ -53,36 +44,4 @@ class ColorLightBase(SoCMini):
             ip_address=str(config.etherbone.ip_address),
             buffer_depth=255,
             data_width=32
-        )
-
-
-class ColorLight_5A_75X(EtherboneBoard):
-    """
-    Configuration for Colorlight 5A-75B and 5A-75E cards:
-
-    """
-    board_type: Literal[
-        '5A-75B v6.1',
-        '5A-75B v7.0',
-        '5A-75B v8.0',
-        '5A-75E v6.0',
-        '5A-75E v7.1'
-    ]
-    ethphy: EthPhy = Field(
-        ...
-    )
-    etherbone: Etherbone = Field(
-        ...,
-    )
-
-    def _generate_soc(self):
-        """Returns the board on which LitexCNC firmware is designed for.
-        """
-        # Split the board and the revision from the board type and create the SoC
-        # based on the board type.
-        board, revision = self.board_type.split('v')
-        return ColorLightBase(
-            board=board.strip().lower(),
-            revision=revision.strip().lower(),
-            config=self
         )
